pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "toonpy"
version = "0.2.0"
description = "Multi-database adapter library that converts query results to TOON format for efficient LLM usage"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Ameya Khot", email = "ameyakhot18@gmail.com"}
]
keywords = ["toon", "database", "adapter", "llm", "token-efficient", "postgresql", "mysql", "mongodb", "token-oriented-object-notation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "python-toon>=1.0.0",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9.0"]
mysql = ["pymysql>=1.0.0"]
mongodb = ["pymongo>=4.0.0"]
all = ["psycopg2-binary>=2.9.0", "pymysql>=1.0.0", "pymongo>=4.0.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/ameyakhot/toondb"
Repository = "https://github.com/ameyakhot/toondb"
Documentation = "https://github.com/ameyakhot/toondb#readme"
Issues = "https://github.com/ameyakhot/toondb/issues"

[tool.setuptools]
packages = ["toonpy"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt"]

